ImageGather.jl documentation

ImageGather.jl documentation

API

Index

API

API reference

Surface offset gathers

ImageGather.double_rtm_cig — Method

double_rtm_cig(model, q, data, offsets, options)

Compute the single shot contribution to the surface offset gather via double rtm. This single source contribution consists of the following steps:

Mute direct arrival in the data.

Compute standard RTM $R$.

Compute the offset RTM $R_o$ for the the weighted data where each trace is weighted by its offset (rec_x - src_x).

Compute the envelope $R_e = \mathcal{E}(R)$ and $R_{oe} = \mathcal{E}(R_o)$.

Compute the offset map $\frac{R_e \odot R_{oe}}{R_e \odot R_e + \epsilon}$.

Apply illumination correction and laplace filter $R_l = \mathcal{D} \Delta R$.

Compute each offset contribution $\mathcal{I}[:, h] = R_l \odot \delta[ha - h]_{tol}$ delta_h.

Return $\mathcal{I}$.

ImageGather.surface_gather — Method

surface_gather(model, q, data; offsets=nothing, options=Options())

Compute the surface offset gathers volume (nx (X ny) X nz X no) for via the double rtm method whith no offsets.

Parameters

model: JUDI Model structure.

q: Source, judiVector.

data: Obeserved data, judiVector.

offsets: List of offsets to compute the gather at. Optional (defaults to 0:10*model.d:model.extent)

options: JUDI Options structure.

Utility functions

ImageGather.delta_h — Method

delta_h(ha, h, tol)

Compute the binary mask where ha is within tol of h.

ImageGather.envelope — Method

envelope(x)

Envelope of a vector or a 1D vector (see DSP hilbert)

ImageGather.envelope — Method

envelope(x)

Envelope of a vector or a 2D matrix. The envelope over the first dimension is taken for a 2D matrix (see DSP hilbert)

ImageGather.laplacian — Method

laplacian(image; hx=1, hy=1)

2D laplacian of an image with grid spacings (hx, hy)

ImageGather.mute! — Method

mute!(shot, offsets;vp=1500, t0=1/10, dt=0.004)

In place direct wave muting of a shot record with water sound speed vp, time sampling dt and firing time t0.

ImageGather.mute — Method

mute(shot, offsets;vp=1500, t0=1/10, dt=0.004)

Direct wave muting of a shot record with water sound speed vp, time sampling dt and firing time t0.

ImageGather.mv_avg_2d — Method

mv_avg_2d(x; k=5)

2D moving average with a square window of width k

Index

ImageGather.delta_h

ImageGather.double_rtm_cig

ImageGather.envelope

ImageGather.envelope

ImageGather.laplacian

ImageGather.mute

ImageGather.mute!

ImageGather.mv_avg_2d

ImageGather.surface_gather